refactor: move implementation outside of `mod.rs` files
